# api/gen/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    An example OpenAPI specification for the Little PM application, referencing path parameters from components.
  title: Little PM API
  version: 1.0.0
servers:
- description: Production server
  url: https://app.little.pm
...
paths:
...
  /projects/{userUUID}:
    get:
      description: List all projects belonging to a specific user.
      operationId: getUserProjects
      parameters:
      - description: ID of the user
        explode: false
        in: path
        name: userUUID
        required: true
        schema:
          $ref: '#/components/schemas/userUUID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProjectIdea'
                type: array
          description: A list of projects
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: User not found
      summary: Get projects for user
      tags:
      - Project Configuration Service
      x-accepts:
      - application/json
      x-tags:
      - tag: Project Configuration Service
    post:
      description: Create a new project idea or configuration for a specific user.
      operationId: createUserProject
      parameters:
      - description: ID of the user
        explode: false
        in: path
        name: userUUID
        required: true
        schema:
          $ref: '#/components/schemas/userUUID'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectIdea'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectIdea'
          description: Project created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: User not found
      summary: Create new project for user
      tags:
      - Project Configuration Service
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Project Configuration Service
  /projects/{userUUID}/{projectUUID}:
    get:
      description: Retrieve details of a specific project for a user.
      operationId: getProject
      parameters:
      - description: ID of the user
        explode: false
        in: path
        name: userUUID
        required: true
        schema:
          $ref: '#/components/schemas/userUUID'
        style: simple
      - description: ID of the project
        explode: false
        in: path
        name: projectUUID
        required: true
        schema:
          $ref: '#/components/schemas/projectUUID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectIdea'
          description: Project found
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Project or user not found
      summary: Get project by UUID
      tags:
      - Project Configuration Service
      x-accepts:
      - application/json
      x-tags:
      - tag: Project Configuration Service
    put:
      description: Update the details of an existing project for a user.
      operationId: updateProject
      parameters:
      - description: ID of the user
        explode: false
        in: path
        name: userUUID
        required: true
        schema:
          $ref: '#/components/schemas/userUUID'
        style: simple
      - description: ID of the project
        explode: false
        in: path
        name: projectUUID
        required: true
        schema:
          $ref: '#/components/schemas/projectUUID'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectIdea'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectIdea'
          description: Project updated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Project or user not found
      summary: Update an existing project
      tags:
      - Project Configuration Service
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Project Configuration Service
...
components:
...
  schemas:
    userUUID:
      description: ID of the user
      example: 665c599d-5c8d-4d20-aaab-7ffaba150606
      format: uuid
      type: string
    projectUUID:
      description: ID of the project
      example: 665c599d-5c8d-4d20-aaab-7ffaba150606
      format: uuid
      type: string
...
    Error:
      description: Error object
      example:
        message: Error message
      properties:
        message:
          description: Error message
          example: Error message
          type: string
      type: object
...
    ProjectIdea:
      example:
        ideaUUID: 665c599d-5c8d-4d20-aaab-7ffaba150606
        teamSize: 6
        requiredInvestment: 0.8008281904610115
        ideaName: ideaName
        productType: productType
      properties:
        ideaUUID:
          description: ID of the project idea
          example: 665c599d-5c8d-4d20-aaab-7ffaba150606
          format: uuid
          type: string
        ideaName:
          type: string
        requiredInvestment:
          type: number
        productType:
          type: string
        teamSize:
          type: integer
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: JWT token
      scheme: bearer
      type: http